--- a/src/regexpr.pas
+++ b/src/regexpr.pas
@@ -914,7 +914,7 @@ uses
 const
   // TRegExpr.VersionMajor/Minor return values of these constants:
   REVersionMajor = 1;
-  REVersionMinor = 171;
+  REVersionMinor = 173;
 
   OpKind_End = REChar(1);
   OpKind_MetaClass = REChar(2);