Many (if not most) other reg ex engines allow patterns to be quantified even if they could be empty or match 0-len.

`(a|\b)*` may match zero len, but should be valid.

For `{}` they even eval the given amount of times: https://regex101.com/r/BvQROU/1 (change the "4" to "5" in the {})

----

Dump() had a potential uninitialised result.

----
```
  RegEx := TRegExpr.Create('((a){2000,4000}){2000,4000}');
//  RegEx.MaxEvalStackDepth := 100000;
  RegEx.InputString := StringOfChar('a', 100000);
  RegEx.Exec();
```
The above would cause a stack overflow.

I added a limit, to avoid the exception.

For my tests on Windows: 
- Win 10 - 64 bit fails at stack depth 45000
- Win 10 - 32 bit fails at stack depth 59000

Obviously stack size can be configured by the user, So numbers can change.

Moving more of the local vars to `TRegExprMatchPrimLocals =   record` and have them share mem, could improve the max depth.

I thought about wrapping `MatchPrim` in an outer function (which would do the inc/dec) and avoid the try/finally.
But that depends on `{$IFDEF InlineFuncs}`, and even then inlining isn't guaranteed, and if it is not inlined it is increasing stack usage.