@User4martin 2023-11-11
Just done some tests...

The try finally causes a big slowdown. So that needs to be skipped for now / or replaced by a better solution