@User4martin 2023-11-11
> The try finally causes a big slowdown. So that needs to be skipped for now / or replaced by a better solution

- `try finally` made my app (which does mostly, but not only regex matching) go from 64 secs to 101 secs. +57%
- Changing each `exit` to `begin dec(matchStackDepth); exit; end;  went from 64 to 74.  +15%
- Comparing   `if InitialFrame - get_frame > fMaxEvalStackDepth then begin` went from 64 to 68. +6%

The last one is not to bad, but I don't know if it works for Delphi. And it depends on the CPU. It needs to know if the stackpointer goes up or down (on intel it goes down, so the current `get_frame` is smaller.)  fMaxEvalStackDepth must then be the size of the stack avail, eg 1 MB or whatever.