@Alexey-T 2023-11-11
even 15% slowdown is bad thing. we add small-needed feature of regex, and perfomance goes down? bad.

your patch is hard to understand  for me (1st of them). so i gueess i just test it and then apply.
if @andgineer can review it, i welcome it.