@User4martin 2023-11-11
> even 15% slowdown is bad thing. we add small-needed feature of regex, and perfomance goes down? bad.

True, 15% is a lot.

But "small-need", well if an editor allows the user to enter search terms, then it can easily end up a stack overflow. For an editor it would not be a surprise to have a longer text (even 100k), and nested loops (just 2 or 3) can quickly reach the limit. 

It may be possible to check for loops only, assuming that other recursions are limited already (IIRC recursions to named captures have a limit, Open-"("-groups too, branches probably too ... Not looked at all others where `MatchPrim` is recursed.

However many of the above can re-occur in each iteration of a loop => meaning that for 1 loop iteration the stack grows by 10 or more frames.

----

Actually just seen, there is an older attempt `if regNestedCalls > MaxRegexBackTracking then` => it is commented though.

----
----


> your patch is hard to understand for me (1st of them). so i gueess i just test it and then apply. if @andgineer can review it, i welcome it.

1) The parser checks for FLAG_HAS_WIDTH, if it may be zero it uses `EmitComplexBraces`. Because this is the only loop that deals with that case

2) Because absence of FLAG_HAS_WIDTH no longer gives an error, I added FLAG_NOT_QUANTIFIABLE => this is set by ^ or \b so ^* or \b* can be rejected. It is not forwarded when it is in `()` so `(\b)*` is allowed. 

3) 'CurrentRegInput' is used to detect if the input moved (if not it is zero width). OP_LOOPENTRY sets it to nil, because it always goes directly to OP_LOOP without processing any input (so it would never advance). OP_LOOP then sets it, and detects it. (For all else it is kept with the loop data, ensuring that for nested loops the correct record is used)

The way nested loops are evaled is tricky (but exists before the patch).<br/>
I.e. if an inner loop tests that the pattern after the loop matches at the point of the current iteration, then that test will run into the OP_LOOP of the outer loop, and therefore also check if the outer loop can get to a valid iteration count (and validating the outer loop may run further instances of the inner loop for each iteration that the outer loop is still taking...)

`Result := (BracesMax = MaxBracesArg) and` => detects `{n,m}` instead of `*` or `+`.  As explained `{n,m}` should run the expected count, even if zero width => because it can changed captures contained in look around. This is the same on he regexp101 site.
However `*` or `+` can not run the 4 million times....