@User4martin 2023-11-11
I've tried a diff approach. Handing the depth in via param https://github.com/User4martin/TRegExpr/commit/0215f19e36bf3e442c20e60d1d9ce8c082550d10

And only checking for the max inside loops (I have to check, but I think the other OP_* can only add limited amount before they must go through a loop. => though OP_BACK may have to be checked, but it probably is never used with captures....

Below are the benchmarks (my own project increased by 5 %)

```
                                                        Time     | Time with stack limit
==============================================================================
                               /Twain/ :          31 ms    31 ms |       16 ms    31 ms
                           /(?i)Twain/ :          47 ms    47 ms |       79 ms    63 ms
                          /[a-z]shing/ :         250 ms   250 ms |      250 ms   234 ms
          /Huck[a-zA-Z]+|Saw[a-zA-Z]+/ :          31 ms    31 ms |       31 ms    31 ms
                           /\b\w+nn\b/ :         250 ms   250 ms |      234 ms   235 ms
                    /[a-q][^u-z]{13}x/ :         641 ms   640 ms |      672 ms   656 ms
         /Tom|Sawyer|Huckleberry|Finn/ :          47 ms    47 ms |       31 ms    32 ms
     /(?i)Tom|Sawyer|Huckleberry|Finn/ :         187 ms   203 ms |      203 ms   203 ms
 /.{0,2}(Tom|Sawyer|Huckleberry|Finn)/ :        2656 ms  2657 ms |     2688 ms  2703 ms     1.5%
 /.{2,4}(Tom|Sawyer|Huckleberry|Finn)/ :        2828 ms  2859 ms |     2906 ms  2875 ms     1.2%
   /Tom.{10,25}river|river.{10,25}Tom/ :          47 ms    62 ms |       47 ms    62 ms
                        /[a-zA-Z]+ing/ :         625 ms   625 ms |      578 ms   579 ms
               /\s[a-zA-Z]{0,12}ing\s/ :         265 ms   282 ms |      266 ms   265 ms
       /([A-Za-z]awyer|[A-Za-z]inn)\s/ :         641 ms   656 ms |      657 ms   641 ms
           /["'][^"']{0,30}[?!\.]["']/ :          78 ms    78 ms |       78 ms    62 ms
```
Though none of them runs very long loops.