@Alexey-T 2023-11-12
>Handing the depth in via param

It looks OK, it does simplify this proposal branch? Commit is out of branch.