@User4martin 2023-11-12
Actually there is a better way. WE can just catch `on E: EStackOverflow do` and set/throw a reg-ex error.

That way usercode does not need to deal with it, but can just do the normal error checking.