@User4martin 2023-11-12
Just wondering...

If a regex exceeds the stack => should that 
- call `Error` and raise an exception
- `Exec`/`Match` return false


At first it seems obvious => raise an exception. But currently all the errors that can happen during eval are internal. The user code will work even if it does not wrap `Exec`/`Match` into a `try except`.

If "too complex" throws an error, then all user code must add exception handling (and that is time consuming)