@Alexey-T 2023-11-12
>If "too complex" throws an error, then all user code must add exception handling (and that is time consuming)

yes, I agree with you, better to avoid changing user code. and return False.