@Alexey-T 2023-11-12
1. strange:
```
  Result := False;
    Result := MatchPrim(regCodeWork);
```

2.    
```
  except
    on E: EStackOverflow do begin
      Result := False;
      //Error(reeLoopStackExceeded);
    end;
  end;
```
"else raise" is missed?