@Alexey-T 2023-11-12
3. speed is slower now, still?