@User4martin 2023-11-12
> 3. speed is slower now, still?

My project has no noticeable change (less than half a second for over a minute runtime => that is normal fluctuation)

Same for the benchmark (here the difference between 2 runs are bigger, than between with/without).

It's one try/except per run. So a single run will not have noticeable diff, however if a user does 100k runs of very simple patterns (like `^1` which would be really fast to run) then the 100k except blocks could add...<br/>
I tried though, on Windows 40m runs of that pattern => no diff within GetTickCount  (1.6 secs) .

**I haven't tried on Linux, different exception handling could mean different outcome.**

> "else raise" is missed?

Ups, but maybe...

It would be good to have an option. `raise` or `just set errorcode`. <br/>
`property RaiseForRuntimeError` (default true)

then this could be for all exceptions.