@User4martin 2023-11-13
> you can 'concatenate' git commits (I dunno how to do it in git) so diff will be smaller.

git REBASE... done.
=> I had the below already written, so it refers to the not-yet-merged. "Current" no longer exists. "NEW" is what the top of the branch is now.

Before the rebase: https://github.com/User4martin/TRegExpr/commits/e34aedb6949a2ea5c9a8f9a2b85539fdf260139b

----


"Current" is before e34aedb6949a2ea5c9a8f9a2b85539fdf260139b
"NEW" is at e34aedb6949a2ea5c9a8f9a2b85539fdf260139b

The NEW code reduces the speed impact of "turning exception to errors", IF (and only IF) it is turned off (property RaiseForRuntimeError := **True**). => That is the original behaviour, in which exceptions are going to the user. <br/>===> Compare "New raise OFF" (behave as original, but has new property) and "No Raise" (original, before any change)


I run the benchmark with `-c 20` (time is avg of 20 runs for each row). For some I repeated the run (again with `-c 20`), and even at that many runs some rows have noticeable fluctutation.

I added new benchmarks, that have a high match count, this means the setup (containing the try/except) is run more often. While other benchmarks may have few matches, but will have to do deep recursions for the loop (hitting code that is at the start of MatchPrim).


- full-check: Current code with RegExpWithStackOverflowCheck defined
- current:   Current code WITHOUT RegExpWithStackOverflowCheck defined
- NO raise:  "simulate original" => raise/except commented out - exception ALWAYS handed to user (porperty RaiseForRuntimeErro removed)<br/>=> stack overflow will not be handled
- NEW raise2Err OFF (true): NEW CODE: - exception ALWAYS handed to user  (property RaiseForRuntimeError := **True**)<br/>=> stack overflow will not be handled (behave as original)
- NEW raise2Err ON (false):  NEW CODE: - exception NOT handed to user  (property RaiseForRuntimeError := **False**)
- NEW raise2Err OFF (true) - full: NEW CODE: - exception ALWAYS handed to user / with RegExpWithStackOverflowCheck defined<br/>This is like only RegExpWithStackOverflowCheck 
- FOR FUN:  same as "NEW raise OFF (true)" but with 1-byte char. (and only -c 5)

```
Windows 10 - 64 bit
                                             Match count  # full-check           # current              # NO raise             # NEW raise2Err OFF  # NEW raise2Err ON # NEW raise2Err OFF - full ## FOR FUN
											                                                              (Original- ish)                                                 (check stack)
=======================================================================================================================================================================================================
                                     /Twain/ :       2388 #     48 ms |    49 ms #     49 ms |    48 ms #    39 ms |    39 ms  #    35 ms |    31 ms  #    31 ms       #     39 ms            ##     19 ms 
                                 /(?i)Twain/ :       2657 #     85 ms |    85 ms #     78 ms |    74 ms #    58 ms |    58 ms  #    56 ms |    56 ms  #    57 ms       #     68 ms            ##     59 ms 
                                /[a-z]shing/ :       1877 #    373 ms |   374 ms #    343 ms |   276 ms #   239 ms |   239 ms  #   238 ms |   241 ms  #   238 ms       #    375 ms            ##    219 ms 
                /Huck[a-zA-Z]+|Saw[a-zA-Z]+/ :        396 #     49 ms |    50 ms #     49 ms |    49 ms #    40 ms |    39 ms  #    35 ms |    32 ms  #    32 ms       #     39 ms            ##     20 ms 
                                         /./ :   18905427 #    682 ms |   682 ms #    459 ms |   531 ms #   468 ms |   470 ms  #   461 ms |   462 ms  #   461 ms       #    666 ms            ##    406 ms 
                                       /(.)/ :   18905427 #    925 ms |  1061 ms #    712 ms |   771 ms #   743 ms |   685 ms  #   728 ms |   741 ms  #   735 ms       #    918 ms            ##    664 ms 
                                         /e/ :    1781425 #    121 ms |   122 ms #    105 ms |   106 ms #    82 ms |    82 ms  #    80 ms |    82 ms  #    82 ms       #    101 ms            ##     67 ms 
                                       /(e)/ :    1781425 #    144 ms |   194 ms #    125 ms |   125 ms #   102 ms |   101 ms  #   103 ms |   103 ms  #   103 ms       #    124 ms            ##     91 ms 
                              /(?s).{1,45} / :     475715 #     41 ms |    65 ms #     38 ms |    37 ms #    37 ms |    38 ms  #    38 ms |    37 ms  #    37 ms       #     43 ms            ##     22 ms 
                            /(?s)\G.{1,45} / :      10616 #    356 ms |   388 ms #    185 ms |   186 ms #   340 ms |   198 ms  #   169 ms |   160 ms  #   164 ms       #    320 ms            ##    167 ms 
                            /\G(?s).{1,45} / :      10616 #     16 ms |    16 ms #     15 ms |    15 ms #    16 ms |    16 ms  #    16 ms |    16 ms  #    17 ms       #     15 ms            ##      0 ms 
                             /(?s).{1,45}? / :    3241534 #    192 ms |   194 ms #    169 ms |   169 ms #   167 ms |   171 ms  #   167 ms |   173 ms  #   182 ms       #    200 ms            ##    152 ms 
                           /(?s)\G.{1,45}? / :      69431 #    360 ms |   359 ms #    189 ms |   189 ms #   176 ms |   177 ms  #   172 ms |   168 ms  #   164 ms       #    321 ms            ##    170 ms 
                           /\G(?s).{1,45}? / :      69431 #     19 ms |    19 ms #     18 ms |    19 ms #    18 ms |    18 ms  #    18 ms |    19 ms  #    19 ms       #     20 ms            ##      3 ms 
                                 /\b\w+nn\b/ :        359 #    390 ms |   390 ms #    277 ms |   277 ms #   250 ms |   275 ms  #   243 ms |   249 ms  #   246 ms       #    371 ms            ##    290 ms 
                          /[a-q][^u-z]{13}x/ :       4929 #    793 ms |   985 ms #    674 ms |   695 ms #   640 ms |   640 ms  #   685 ms |   711 ms  #   675 ms       #    785 ms            ##    648 ms 
               /Tom|Sawyer|Huckleberry|Finn/ :       3015 #     54 ms |    53 ms #     53 ms |    53 ms #    43 ms |    43 ms  #    38 ms |    35 ms  #    35 ms       #     44 ms            ##     24 ms 
           /(?i)Tom|Sawyer|Huckleberry|Finn/ :       4820 #    263 ms |   257 ms #    230 ms |   230 ms #   200 ms |   214 ms  #   194 ms |   198 ms  #   197 ms       #    219 ms            ##    352 ms 
       /.{0,2}(Tom|Sawyer|Huckleberry|Finn)/ :       3015 #   2992 ms |  3196 ms #   2792 ms |  2839 ms #  2697 ms |  2685 ms  #  2713 ms |  2722 ms  #  2719 ms       #   2878 ms            ##   2551 ms 
       /.{2,4}(Tom|Sawyer|Huckleberry|Finn)/ :       2220 #   3405 ms |  3299 ms #   3071 ms |  3191 ms #  2937 ms |  2917 ms  #  2996 ms |  2991 ms  #  3046 ms       #   3122 ms            ##   2745 ms 
         /Tom.{10,25}river|river.{10,25}Tom/ :          2 #     82 ms |    82 ms #     73 ms |    92 ms #    60 ms |    60 ms  #    54 ms |    57 ms  #    70 ms       #     66 ms            ##     43 ms 
                              /[a-zA-Z]+ing/ :      95863 #    750 ms |   748 ms #    609 ms |   676 ms #   625 ms |   628 ms  #   591 ms |   628 ms  #   657 ms       #    793 ms            ##    600 ms 
                     /\s[a-zA-Z]{0,12}ing\s/ :      67810 #    318 ms |   332 ms #    287 ms |   286 ms #   278 ms |   278 ms  #   268 ms |   271 ms  #   271 ms       #    348 ms            ##    261 ms 
             /([A-Za-z]awyer|[A-Za-z]inn)\s/ :        313 #    825 ms |   825 ms #    696 ms |   694 ms #   658 ms |   642 ms  #   646 ms |   657 ms  #   659 ms       #    876 ms            ##    614 ms 
                 /["'][^"']{0,30}[?!\.]["']/ :       9857 #     95 ms |    94 ms #     88 ms |    86 ms #    85 ms |    83 ms  #    79 ms |    73 ms  #    74 ms       #     83 ms            ##     60 ms 
                   /Tom(.{3,3}|.{5,5})*Finn/ :         11 #   3683 ms |  3562 ms #   3146 ms |  3403 ms #  3127 ms |  3293 ms  #  3140 ms |  3167 ms  #  3242 ms       #   3387 ms            ##   3132 ms 
                       /Tom(...|.....)*Finn/ :         11 #   1891 ms |  2124 ms #   1708 ms |  1802 ms #  1739 ms |  1671 ms  #  1785 ms |  1874 ms  #  1979 ms       #   1754 ms            ##   1564 ms 
                      /Tom(...|.....)*?Finn/ :         11 #   1927 ms |  1763 ms #   1651 ms |  1714 ms #  1633 ms |  1754 ms  #  1641 ms |  1646 ms  #  1673 ms       #   1721 ms            ##   1537 ms 
         /Tom((...|.....){2,9}\s){1,5}?Finn/ :         11 #   4865 ms |  4573 ms #   4392 ms |  4355 ms #  4305 ms |  4596 ms  #  4391 ms |  4467 ms  #  4314 ms       #   4517 ms            ##   4228 ms 
         /Tom((...|.....){2,9}?\s){1,5}Finn/ :         11 #   5181 ms |  4576 ms #   4309 ms |  4289 ms #  4245 ms |  4282 ms  #  4310 ms |  4321 ms  #  4283 ms       #   4541 ms            ##   4172 ms 
                           /\G(?is).(?=.*$)/ :   20045118 #   1212 ms |  1186 ms #    986 ms |   982 ms #   943 ms |   921 ms  #   968 ms |   935 ms  #   932 ms       #   1203 ms            ##   1014 ms 
                   /\G(?is).(?=(.){1,5}?$)?/ :   20045118 #   4123 ms |  4063 ms #   3541 ms |  3532 ms #  3479 ms |  3605 ms  #  3869 ms |  3568 ms  #  3728 ms       #   4028 ms            ##   3592 ms 
                          /\G(?is).(?=.*+$)/ :   20045118 #   1182 ms |  1178 ms #   1096 ms |   975 ms #   927 ms |   889 ms  #   963 ms |   953 ms  #   952 ms       #   1283 ms            ##    995 ms 
   /\G(?is).{10,10}(?=(e|y|on|fin|.){0,20})/ :    2004511 #   2557 ms |  2560 ms #   2395 ms |  2433 ms #  2427 ms |  2411 ms  #  2392 ms |  2417 ms  #  2421 ms       #   2546 ms            ##   3174 ms 
 /\G(?is).{10,10}(?=(?>e|y|on|fin|.){0,20})/ :    2004511 #   2573 ms |  2554 ms #   2395 ms |  2460 ms #  2432 ms |  2392 ms  #  2396 ms |  2415 ms  #  2461 ms       #   2541 ms            ##   3160 ms 
                             /Tom(?!.*Finn)/ :       2441 #     50 ms |    50 ms #     49 ms |    49 ms #    43 ms |    41 ms  #    35 ms |    32 ms  #    34 ms       #     40 ms            ##     26 ms 
Total                                                        42622 ms | 42108 ms #  37052 ms | 37708 ms # 36298 ms | 36651 ms  # 36713 ms | 36708 ms  # 37004 ms       #  40412 ms            ##  36856 ms 
```

Linux benchmares with `-c 5`
```
Linux 64 bit (virtualbox)

                                              | Match count # ORIGINAL    # NEW raise2Err OFF # NEW raise2Err ON # NEW raise2Err OFF - full
											              (before any change)                                       (check stack)
============================================================================================================================================================
regexpr.pas:
                                        /Twain/ :      2388 #       41 ms #      42 ms        #      41 ms       #      46 ms
                                    /(?i)Twain/ :      2657 #       94 ms #      97 ms        #      99 ms       #      98 ms
                                   /[a-z]shing/ :      1877 #      944 ms #     958 ms        #     976 ms       #     948 ms
                   /Huck[a-zA-Z]+|Saw[a-zA-Z]+/ :       396 #       42 ms #      44 ms        #      46 ms       #      48 ms
                                            /./ :  18905425 #     1376 ms #    1432 ms        #    1731 ms       #    1411 ms
                                          /(.)/ :  18905425 #     2203 ms #    2313 ms        #    2600 ms       #    2328 ms
                                            /e/ :   1781425 #      177 ms #     187 ms        #     212 ms       #     183 ms
                                          /(e)/ :   1781425 #      255 ms #     264 ms        #     293 ms       #     264 ms
                                 /(?s).{1,45} / :    475715 #       83 ms #      86 ms        #      93 ms       #      86 ms
                               /(?s)\G.{1,45} / :     10616 #      767 ms #     786 ms        #     800 ms       #     814 ms
                               /\G(?s).{1,45} / :     10616 #       26 ms #      25 ms        #      28 ms       #      26 ms
                                /(?s).{1,45}? / :   3241534 #      413 ms #     437 ms        #     490 ms       #     447 ms
                              /(?s)\G.{1,45}? / :     69431 #      765 ms #     810 ms        #     813 ms       #     810 ms
                              /\G(?s).{1,45}? / :     69431 #       32 ms #      34 ms        #      34 ms       #      33 ms
                                    /\b\w+nn\b/ :       359 #      551 ms #     584 ms        #     589 ms       #     568 ms
                             /[a-q][^u-z]{13}x/ :      4929 #     1284 ms #    1327 ms        #    1299 ms       #    1292 ms
                  /Tom|Sawyer|Huckleberry|Finn/ :      3015 #       54 ms #      55 ms        #      55 ms       #      59 ms
              /(?i)Tom|Sawyer|Huckleberry|Finn/ :      4820 #      460 ms #     475 ms        #     468 ms       #     466 ms
          /.{0,2}(Tom|Sawyer|Huckleberry|Finn)/ :      3015 #     8704 ms #    8974 ms        #    8718 ms       #    8808 ms
          /.{2,4}(Tom|Sawyer|Huckleberry|Finn)/ :      2220 #     9448 ms #    9702 ms        #    9478 ms       #    9627 ms
            /Tom.{10,25}river|river.{10,25}Tom/ :         2 #      105 ms #     109 ms        #     106 ms       #     106 ms
                                 /[a-zA-Z]+ing/ :     95863 #     2501 ms #    2580 ms        #    2492 ms       #    2544 ms
                        /\s[a-zA-Z]{0,12}ing\s/ :     67810 #      919 ms #     953 ms        #     932 ms       #     938 ms
                /([A-Za-z]awyer|[A-Za-z]inn)\s/ :       313 #     2893 ms #    2998 ms        #    2916 ms       #    2952 ms
                    /["'][^"']{0,30}[?!\.]["']/ :      9857 #      126 ms #     125 ms        #     125 ms       #     132 ms
                      /Tom(.{3,3}|.{5,5})*Finn/ :        11 #    10526 ms #   11186 ms        #   11149 ms       #   10902 ms
                          /Tom(...|.....)*Finn/ :        11 #     5651 ms #    6038 ms        #    6130 ms       #    5952 ms
                         /Tom(...|.....)*?Finn/ :        11 #     4934 ms #    5162 ms        #    5285 ms       #    5053 ms
            /Tom((...|.....){2,9}\s){1,5}?Finn/ :        11 #    13297 ms #   14304 ms        #   14221 ms       #   13414 ms
            /Tom((...|.....){2,9}?\s){1,5}Finn/ :        11 #    13365 ms #   14266 ms        #   14142 ms       #   13553 ms
                              /\G(?is).(?=.*$)/ :  20045118 #     3413 ms #    3465 ms        #    3792 ms       #    3492 ms
                      /\G(?is).(?=(.){1,5}?$)?/ :  20045118 #    12193 ms #   12457 ms        #   12945 ms       #   12432 ms
                             /\G(?is).(?=.*+$)/ :  20045118 #     3405 ms #    3518 ms        #    3757 ms       #    3470 ms
      /\G(?is).{10,10}(?=(e|y|on|fin|.){0,20})/ :   2004511 #     7094 ms #    7081 ms        #    7156 ms       #    6958 ms
    /\G(?is).{10,10}(?=(?>e|y|on|fin|.){0,20})/ :   2004511 #     7123 ms #    7146 ms        #    7159 ms       #    6989 ms
                                /Tom(?!.*Finn)/ :      2441 #       45 ms #      47 ms        #      50 ms       #      53 ms
Total:                                                          115323 ms #  120079 ms        #  121237 ms       #  117316 ms
```

So much for fluctuations, the last column is faster than the 2nd column => yet it does definetily more work. 

But that can actually happen. It is possible that the inserted code moved the other code inside the function, and the same code, but moved a few bytes can be faster (some dependency on 32byte boundaries).